ExtractYearFromDate .task Envi Node
custom .task / .pro files to add an envi modeling node that pulls the first 4 digit number from a string integer and converts it to an integer. This allows a year-based number to append while keeping an identifier prefix in iterated datasets. 

Built for landsat 08/09 data but can be converted to any data with a fixed naming structure.